add colaboration

+++ src/api/playlistSong/playlistSongHandler.js
@@ -1,8 +1,10 @@
 const { nanoid } = require('nanoid');
 const pool = require('../../database');
+const tokenManager = require('../../tokenize/TokenManager');
 const OpenMusicErrorHandling = require('../../exception/OpenMusicErrorHandling');
 const { PostPlaylistSongPayloadSchema } = require('../../validator/playlistScheme');
 const songHelper = require('../song/helper');
+const playlistSongHelper = require('./helper');
 
 const handler = {
   async postPlaylistSongHandler(req, h) {
@@ -15,7 +17,11 @@ const handler = {
       const { playlistId } = req.params;
       const { songId } = req.payload;
       const id = `playlistSong-${nanoid(16)}`;
+      const { authorization } = req.headers;
+      const token = authorization.replace(/^Bearer\s*/, '');
+      const { id: owner } = tokenManager.verifyAccessToken(token);
 
+      playlistSongHelper.validateOwner(playlistId, owner);
       songHelper.validateSongById(songId);
 
       await pool.query(`INSERT INTO playlistsongs(id, playlist_id, song_id) VALUES('${id}', '${playlistId}', '${songId}')`);
@@ -34,6 +40,11 @@ const handler = {
   async getPlaylistSongHandler(req, h) {
     try {
       const { playlistId } = req.params;
+      const { authorization } = req.headers;
+      const token = authorization.replace(/^Bearer\s*/, '');
+      const { id: owner } = tokenManager.verifyAccessToken(token);
+
+      playlistSongHelper.validateOwner(playlistId, owner);
 
       const result = await pool.query(`
       SELECT songs.id, songs.title, songs.performer
@@ -41,7 +52,7 @@ const handler = {
       JOIN playlistsongs
       ON songs.id = playlistsongs.song_id AND playlistsongs.id = '${playlistId}'`);
       if (result.rows.length === 0) {
-        throw OpenMusicErrorHandling('Lagu tidak ditemukan dalam playlist', 404);
+        throw OpenMusicErrorHandling('Data Not Found', 404);
       }
       const response = h.response({
         status: 'success',
@@ -59,7 +70,11 @@ const handler = {
     try {
       const { playlistId } = req.params;
       const { songId } = req.payload;
+      const { authorization } = req.headers;
+      const token = authorization.replace(/^Bearer\s*/, '');
+      const { id: owner } = tokenManager.verifyAccessToken(token);
 
+      playlistSongHelper.validateOwner(playlistId, owner);
       songHelper.validateSongById(songId);
 
       await pool.query(`DELETE FROM playlistsongs WHERE id = '${playlistId}' AND song_id = '${songId}'`);
